Remove single-chat team upgrade path
