+++ packages/desktop/src/renderer/pages/conversation/platforms/acp/AcpSendBox.tsx
@@ -409,7 +409,6 @@ Please check your local CLI tool authentication status`,
     enabled: true,
     isBusy,
     runtimeGate: commandQueueRuntimeGate,
-    teamUpgradeHandoffReady: Boolean(teamRuntime && teamSendMessage),
     onExecute: executeCommand,
   });
