+++ packages/desktop/src/renderer/utils/emitter.ts
@@ -32,7 +32,6 @@ interface EventTypes {
   'chat.history.refresh': void;
   // 会话删除事件 / Conversation deletion event
   'conversation.deleted': [string]; // conversation_id
-  'conversation.commandQueue.deferAfterTeamUpgrade': [{ conversation_id: string; team_id: string }];
   // 预览面板事件 / Preview panel events
   'preview.open': [
     { content: string; contentType: PreviewContentType; metadata?: { title?: string; file_name?: string } },
