+++ tests/unit/conversation/runtime/conversationCommandQueueDrain.dom.test.tsx
@@ -1,0 +1,147 @@
+/**
+ * @license
+ * Copyright 2025 AionUi (aionui.com)
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { act, renderHook, waitFor } from '@testing-library/react';
+import { createElement, type PropsWithChildren } from 'react';
+import { SWRConfig } from 'swr';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import {
+  type ConversationCommandQueueRuntimeGate,
+  useConversationCommandQueue,
+} from '@/renderer/pages/conversation/platforms/useConversationCommandQueue';
+
+vi.mock('react-i18next', () => ({
+  useTranslation: () => ({
+    t: (_key: string, options?: { defaultValue?: string }) => options?.defaultValue ?? _key,
+  }),
+}));
+
+vi.mock('@arco-design/web-react', () => ({
+  Message: {
+    info: vi.fn(),
+    warning: vi.fn(),
+  },
+}));
+
+const createSwrWrapper = () => {
+  const cache = new Map();
+
+  return function SwrTestWrapper({ children }: PropsWithChildren) {
+    return createElement(
+      SWRConfig,
+      {
+        value: {
+          provider: () => cache,
+          dedupingInterval: 0,
+          revalidateOnFocus: false,
+          revalidateOnReconnect: false,
+        },
+      },
+      children
+    );
+  };
+};
+
+const processingGate: ConversationCommandQueueRuntimeGate = {
+  hydrated: true,
+  canSendMessage: true,
+  isProcessing: true,
+};
+
+const idleGate: ConversationCommandQueueRuntimeGate = {
+  hydrated: true,
+  canSendMessage: true,
+  isProcessing: false,
+};
+
+const storageKey = (conversationId: string) => `conversation-command-queue/${conversationId}`;
+
+const renderQueue = ({
+  conversation_id,
+  runtimeGate,
+  isBusy = false,
+  onExecute = vi.fn().mockResolvedValue(undefined),
+}: {
+  conversation_id: string;
+  runtimeGate: ConversationCommandQueueRuntimeGate;
+  isBusy?: boolean;
+  onExecute?: (item: Parameters<Parameters<typeof useConversationCommandQueue>[0]['onExecute']>[0]) => Promise<void>;
+}) =>
+  renderHook(
+    ({ gate, busy }) =>
+      useConversationCommandQueue({
+        conversation_id,
+        enabled: true,
+        isBusy: busy,
+        runtimeGate: gate,
+        onExecute,
+      }),
+    {
+      initialProps: { gate: runtimeGate, busy: isBusy },
+      wrapper: createSwrWrapper(),
+    }
+  );
+
+describe('useConversationCommandQueue drain', () => {
+  beforeEach(() => {
+    sessionStorage.clear();
+    vi.spyOn(console, 'info').mockImplementation(() => {});
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+    sessionStorage.clear();
+  });
+
+  it('drains a queued command when the runtime becomes idle', async () => {
+    const onExecute = vi.fn().mockResolvedValue(undefined);
+    const { result, rerender } = renderQueue({
+      conversation_id: 'conv-1',
+      runtimeGate: processingGate,
+      onExecute,
+    });
+
+    act(() => {
+      result.current.enqueue({ input: 'queued follow-up', files: [] });
+    });
+    await waitFor(() => expect(result.current.items).toHaveLength(1));
+
+    rerender({ gate: idleGate, busy: false });
+
+    await waitFor(() => expect(onExecute).toHaveBeenCalledTimes(1));
+    expect(onExecute).toHaveBeenCalledWith(expect.objectContaining({ input: 'queued follow-up' }));
+  });
+
+  it('ignores legacy persisted team-upgrade handoff state and drains normally', async () => {
+    const onExecute = vi.fn().mockResolvedValue(undefined);
+    const legacyHandoffKey = ['deferred', 'AfterTeamUpgrade'].join('');
+    sessionStorage.setItem(
+      storageKey('conv-legacy'),
+      JSON.stringify({
+        items: [
+          {
+            id: 'queued-1',
+            input: 'legacy persisted follow-up',
+            files: [],
+            created_at: 1,
+          },
+        ],
+        isPaused: false,
+        [legacyHandoffKey]: true,
+      })
+    );
+
+    renderQueue({
+      conversation_id: 'conv-legacy',
+      runtimeGate: idleGate,
+      onExecute,
+    });
+
+    await waitFor(() => expect(onExecute).toHaveBeenCalledTimes(1));
+    expect(onExecute).toHaveBeenCalledWith(expect.objectContaining({ input: 'legacy persisted follow-up' }));
+    await waitFor(() => expect(sessionStorage.getItem(storageKey('conv-legacy'))).toBeNull());
+  });
+});
